Module tk_http::client

The HTTP/1.x client protocol implementation

Modules

buffered
Simple to use wrappers for dealing with fully buffered requests

Structs

Config
Fine-grained configuration of the HTTP connection

Encoder
This a request writer that you receive in Codec

EncoderDone
This structure returned from Encoder::done and works as a continuation that should be returned from the future that writes request.

Error
HTTP client error

Head
A borrowed structure that represents response headers

Proto
A low-level HTTP/1.x client protocol handler

RecvMode
This type is returned from headers_received handler of either client client or server protocol handler

Traits

Client
A marker trait that applies to a Sink that is essentially a HTTP client

Codec
This is a low-level interface to the http client