Module tk_http::server

HTTP server protocol implementation

Modules

buffered
Higher-level interface for serving fully buffered requests

Structs

Config
Fine-grained configuration of the HTTP server

Encoder
This a response writer that you receive in Codec

EncoderDone
This structure returned from Encoder::done and works as a continuation that should be returned from the future that writes request.

Error
HTTP server error

FutureRawBody
A future that yields RawBody after buffer is empty

Head
A borrowed structure that represents request headers

HeaderIter
Iterator over all meaningful headers for the request

Proto
A low-level HTTP/1.x server protocol handler

RawBody
The actual raw body

RecvMode
This type is returned from headers_received handler of either client client or server protocol handler

WebsocketHandshake
Contains all the imporant parts of a websocket handshake

Enums

RequestTarget
A middle part of the request line

Traits

Codec
The type represents a consumer of a single request and yields a writer of a response (the latter is a ResponseFuture

Dispatcher
This is a low-level interface to the http server