Module tk_http::server::buffered

Higher-level interface for serving fully buffered requests

Structs

BufferedCodec
A codec counterpart of the BufferedDispatcher, might be used with your own dispatcher too

BufferedDispatcher
A dispatcher that allows to process request and return response using a one single function

Request
Buffered request struct

WebsocketFactory
A helper to create a simple websocket (and HTTP) service

WebsocketService
An instance of websocket factory, created by WebsocketFactory itself

Traits

NewService
A trait that you must implement to reply on requests, usually a function

Service
An instance of a NewService for a single request, usually just a function