Struct tk_http::server::RawBody

pub struct RawBody<S> { /* fields omitted */ }

The actual raw body

The object is used to write some data directly to the socket without any buffering/copying. Note that chunked encoding must be handled manually in this case.

This is a tiny wrapper around WriteRaw which is basically tiny wrapper around TcpStream or whatever S represents. Wrappers are used to reconstruct original object, EncoderDone in this case.

Methods

impl<S> RawBody<S>

fn done(self) -> EncoderDone<S>

Returns EncoderDone object that might be passed back to the HTTP protocol