Struct tk_http::websocket::client::SimpleAuthorizer

pub struct SimpleAuthorizer { /* fields omitted */ }

Default handshake handler, if you just want to get websocket connected

Methods

impl SimpleAuthorizer

fn new<A, B>(host: A, path: B) -> SimpleAuthorizer where
A: Into<String>,
B: Into<String>,

Create a new authorizer that sends specified host and path