Struct tk_http::websocket::Accept

pub struct Accept(_);

The Sec-WebSocket-Accept header value

You can add it using enc.format_header("Sec-WebSocket-Accept", accept). Or use any other thing that supports Display.

Methods

impl Accept

fn from_key_bytes(key: &[u8]) -> Accept

Create an Accept header value from a key received in header

Note: key here is a key as passed in header value (base64-encoded) despite that it's accepted as bytes (not as 16 bytes stored in Key)

Note 2: this does not validate a key (which is not required by spec)