Struct tk_http::client::Proto

pub struct Proto<S: Io, C: Codec<S>> { /* fields omitted */ }

A low-level HTTP/1.x client protocol handler

Note, most of the time you need some reconnection facility and/or connection pooling on top of this interface

Methods

impl<S: Io, C: Codec<S>> Proto<S, C>

fn new(conn: S, handle: &Handle, cfg: &Arc<Config>) -> Proto<S, C>

Create a new protocol implementation from a TCP connection and a config

You should use this protocol as a Sink

impl<C: Codec<TcpStream>> Proto<TcpStream, C>

fn connect_tcp(addr: SocketAddr,
cfg: &Arc<Config>,
handle: &Handle)
-> Box<Future<Item=Self, Error=Error>>

A convenience method to establish connection and create a protocol instance